requirements.txt
kafka-python
fastavro